hawk_eye_notify/requirements.txt
Jinja2
pyinotify
docopt
pyyaml
